# openapi_agent_api.yaml
openapi: 3.0.3
info:
  title: Agent Transfer Standard (ATS) — IMO-to-Carrier MVP API
  version: 0.1.0
  description: |
    Minimal carrier-hosted API for Agent/IMO-initiated transfers.
    Scope: Create a transfer, check status, cancel, and (optionally) list.
servers:
  - url: https://api.carrier.example.com
...
paths:
  /ats/transfers:
    get:
      tags: [Transfers]
      summary: List transfers (basic filter)
      operationId: listTransfers
      parameters:
        - in: query
          name: npn
          schema: { type: string }
          description: Optional agent NPN to filter results
        - in: query
          name: state
          schema:
            $ref: "#/components/schemas/TransferState"
          description: Optional state filter
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Paged list (MVP returns a simple array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TransferSummary" }

    post:
      tags: [Transfers]
      summary: Create a new agent transfer
      operationId: createTransfer
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          description: Idempotency key to safely retry POST without duplicates
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferCreateRequest" }
            examples:
              minimal:
                summary: Minimal creation
                value:
                  agent:
                    { npn: "17439285", firstName: "Jordan", lastName: "Lee" }
                  releasingImo: { fein: "12-3456789", name: "Acme IMO" }
                  receivingImo: { fein: "98-7654321", name: "Summit IMO" }
                  effectiveDate: "2026-03-01"
                  consent: { agentAttestation: true, eSignatureRef: "esig_123" }
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL of the newly created transfer resource
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransferCreateResponse" }
              examples:
                created:
                  value: { id: "tr_abc123", state: "SUBMITTED" }
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Conflict (e.g., duplicate active transfer for this agent+carrier)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
...
    TransferCreateRequest:
      type: object
      required: [agent, releasingImo, receivingImo, effectiveDate, consent]
      properties:
        agent:
          $ref: "#/components/schemas/AgentRef"
        releasingImo:
          $ref: "#/components/schemas/OrgRef"
        receivingImo:
          $ref: "#/components/schemas/OrgRef"
        effectiveDate:
          type: string
          format: date
          description: Target date for hierarchy/commission realignment
        consent:
          type: object
          required: [agentAttestation]
          properties:
            agentAttestation:
              type: boolean
              description: True if the agent has authorized this transfer
            eSignatureRef:
              type: string
              description: Reference to an e-signature artifact (opaque to carrier)
        notes:
          type: string
          maxLength: 2000
          description: Optional free text for carrier processing
...
    TransferCreateResponse:
      type: object
      required: [id, state]
      properties:
        id: { type: string, description: Server-assigned transfer id }
        state: { $ref: "#/components/schemas/TransferState" }
    TransferSummary:
      type: object
      required: [id, state, agent]
      properties:
        id: { type: string }
        state: { $ref: "#/components/schemas/TransferState" }
        agent: { $ref: "#/components/schemas/AgentRef" }
        effectiveDate: { type: string, format: date }
...
    TransferState:
      type: string
      enum: [SUBMITTED, VALIDATION, PROCESSING, COMPLETED, REJECTED, WITHDRAWN]
      description: |
        MVP state machine:
          SUBMITTED → VALIDATION → PROCESSING → COMPLETED
                                  └───────────→ REJECTED
        WITHDRAWN indicates an IMO/Agent-initiated cancel.
    AgentRef:
      type: object
      required: [npn]
      properties:
        npn:
          type: string
          description: National Producer Number (unique industry identifier)
        firstName: { type: string }
        lastName: { type: string }
    OrgRef:
      type: object
      required: [fein, name]
      properties:
        fein:
          type: string
          description: Business FEIN for the IMO
        name:
          type: string
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }

    # ===== Status =====
...
security:
  # Placeholder; for MVP demos you can disable auth.
  # In production, use OAuth2 Client Credentials or AWS SigV4/mTLS between IMO and carrier.
  - {}